refactor: expose jsonutil
This package should be exposed to allow users to correctly unmarshal responses by subscriptions.
