+++ pkg/jsonutil/benchmark_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/hasura/go-graphql-client/internal/jsonutil"
+	"github.com/hasura/go-graphql-client/pkg/jsonutil"
 )
 
 func TestUnmarshalGraphQL_benchmark(t *testing.T) {
